Process editor refactoring
