+++ projects/ama-sdk/src/lib/process-editor/process-editor.actions.ts
@@ -21,5 +21,9 @@ import { ServiceParameterMappings } from '../api/types';
 export const UPDATE_SERVICE_PARAMETERS = '[ProcessEditor] Update Service Parameters';
 export class UpdateServiceParametersAction implements Action {
     readonly type = UPDATE_SERVICE_PARAMETERS;
-    constructor(public serviceId: string, public serviceParameterMappings: ServiceParameterMappings) {}
+    constructor(
+        public processId: string,
+        public serviceId: string,
+        public serviceParameterMappings: ServiceParameterMappings
+    ) {}
 }
