+++ src/app/application-editor/services/application-editor.service.ts
@@ -16,7 +16,7 @@
  */
 
 import { Injectable } from '@angular/core';
-import { Application, Process, AmaApi, EntityDialogForm, UploadFileAttemptPayload } from 'ama-sdk';
+import { Application, AmaApi } from 'ama-sdk';
 import { Observable } from 'rxjs';
 
 @Injectable()
@@ -25,22 +25,6 @@ export class ApplicationEditorService {
 
     fetchApplication(applicationId: string): Observable<Partial<Application>> {
         return this.amaApi.Application.get(applicationId);
-    }
-
-    fetchProcesses(applicationId: string): Observable<Process[]> {
-        return this.amaApi.Process.getList(applicationId);
-    }
-
-    createProcess(form: Partial<EntityDialogForm>, applicationId: string): Observable<Process> {
-        return this.amaApi.Process.create(form, applicationId);
-    }
-
-    deleteProcess(processId: string): Observable<any> {
-        return this.amaApi.Process.delete(processId);
-    }
-
-    uploadProcess(payload: UploadFileAttemptPayload): Observable<Process> {
-        return this.amaApi.Process.import(payload.file, payload.applicationId);
     }
 
     exportApplication(applicationId: string): Observable<Blob> {
