+++ src/app/application-editor/store/reducers/application-tree.reducer.spec.ts
@@ -15,12 +15,9 @@
  * limitations under the License.
  */
 
-import { ApplicationTreeState, INITIAL_APPLICATION_TREE_STATE, Process, OpenFilterAction } from 'ama-sdk';
+import { ApplicationTreeState, INITIAL_APPLICATION_TREE_STATE, OpenFilterAction, CONNECTOR, PROCESS } from 'ama-sdk';
 import { applicationTreeReducer } from './application-tree.reducer';
-import { SELECT_APPLICATION, CloseFilterAction } from '../actions/application';
-import { CreateProcessSuccessAction, DeleteProcessSuccessAction } from '../actions/processes';
-import { CONNECTOR, PROCESS } from 'ama-sdk';
-import { UpdateProcessSuccessAction } from '../../../process-editor/store/process-editor.actions';
+import { SELECT_APPLICATION, CloseFilterAction } from '../application-editor.actions';
 
 describe('ApplicationTreeReducer', () => {
     let initState: ApplicationTreeState;
@@ -30,44 +27,6 @@ describe('ApplicationTreeReducer', () => {
     it('should handle SELECT_APPLICATION', () => {
         const newState = applicationTreeReducer(initState, {type: SELECT_APPLICATION});
         expect(newState).toEqual(initState);
-    });
-
-    it ('should handle CREATE_PROCESS_SUCCES', () => {
-        const process = <Process>{
-            id: 'id1',
-            name: 'name'
-        };
-
-        const newState = applicationTreeReducer(initState, new CreateProcessSuccessAction(process));
-
-        expect(newState.processes[process.id]).toEqual(process);
-    });
-
-    it ('should handle UPDATE_PROCESS_SUCCESS', () => {
-        const updateProcessPayload = {
-            processId: 'id1',
-            content: '',
-            metadata: {
-                name: 'new name',
-                description: ''
-            }
-        };
-
-        const process = <Process>{
-            id: 'id1',
-            name: 'name'
-        };
-
-        const state = applicationTreeReducer(initState, new CreateProcessSuccessAction(process));
-        const newState = applicationTreeReducer(state, new UpdateProcessSuccessAction(updateProcessPayload));
-
-        expect(newState.processes[updateProcessPayload.processId].name).toEqual(updateProcessPayload.metadata.name);
-    });
-
-    it ('should handle DELETE_PROCESS_SUCCESS', () => {
-        const newState = applicationTreeReducer(initState, new DeleteProcessSuccessAction('processId'));
-
-        expect(newState.processes['processId']).toBeUndefined();
     });
 
     it('should handle OPEN_FILTER', () => {
