+++ src/app/application-editor/store/reducers/application-tree.reducer.ts
@@ -16,19 +16,8 @@
  */
 
 import { Action } from '@ngrx/store';
-import {
-    GET_PROCESSES_SUCCESS,
-    GetProcessesSuccessAction,
-    CREATE_PROCESS_SUCCESS,
-    CreateProcessSuccessAction,
-    DELETE_PROCESS_SUCCESS,
-    DeleteProcessSuccessAction
-} from '../actions/processes';
 import { ApplicationTreeState, INITIAL_APPLICATION_TREE_STATE as init, OPEN_FILTER, OpenFilterAction } from 'ama-sdk';
-import { Process } from 'ama-sdk';
-import { SELECT_APPLICATION, CLOSE_FILTER, CloseFilterAction } from '../actions/application';
-import { UPDATE_PROCESS_SUCCESS, UpdateProcessSuccessAction } from '../../../process-editor/store/process-editor.actions';
-const keyBy = require('lodash/keyBy');
+import { SELECT_APPLICATION, CLOSE_FILTER, CloseFilterAction } from '../application-editor.actions';
 
 
 export function applicationTreeReducer(state: ApplicationTreeState = init, action: Action): ApplicationTreeState {
@@ -36,23 +25,7 @@ export function applicationTreeReducer(state: ApplicationTreeState = init, actio
 
     switch (action.type) {
         case SELECT_APPLICATION:
-            newState = initTree(state);
-            break;
-
-        case GET_PROCESSES_SUCCESS:
-            newState = setProcesses(state, <GetProcessesSuccessAction>action);
-            break;
-
-        case CREATE_PROCESS_SUCCESS:
-            newState = createProcess(state, <CreateProcessSuccessAction>action);
-            break;
-
-        case UPDATE_PROCESS_SUCCESS:
-            newState = updateProcess(state, <UpdateProcessSuccessAction>action);
-            break;
-
-        case DELETE_PROCESS_SUCCESS:
-            newState = deleteProcess(state, <DeleteProcessSuccessAction>action);
+            newState = { ...init };
             break;
 
         case OPEN_FILTER:
@@ -66,52 +39,6 @@ export function applicationTreeReducer(state: ApplicationTreeState = init, actio
         default:
             newState = Object.assign({}, state);
     }
-
-    return newState;
-}
-
-function initTree(state: ApplicationTreeState): ApplicationTreeState {
-    return { ...init };
-}
-
-
-function setProcesses(state: ApplicationTreeState, action: GetProcessesSuccessAction): ApplicationTreeState {
-    const newState = { ...state };
-    newState.processes = keyBy(action.processes, 'id');
-    return newState;
-}
-
-function createProcess(state: ApplicationTreeState, action: CreateProcessSuccessAction): ApplicationTreeState {
-    const newState = { ...state },
-        newProcess = action.process;
-
-    newState.processes = {
-        ...state.processes,
-        [newProcess.id]: newProcess
-    };
-
-    return newState;
-}
-
-function updateProcess(state: ApplicationTreeState, action: UpdateProcessSuccessAction): ApplicationTreeState {
-    const newState = { ...state },
-        oldProcess = state.processes[action.payload.processId],
-        newProcess = { ...oldProcess, ...action.payload.metadata };
-
-    newState.processes = {
-        ...state.processes,
-        [action.payload.processId]: <Process>newProcess
-    };
-
-    return newState;
-}
-
-function deleteProcess(state: ApplicationTreeState, action: DeleteProcessSuccessAction): ApplicationTreeState {
-    const newState = { ...state };
-    const processId = action.processId;
-
-    newState.processes = { ...state.processes };
-    delete newState.processes[processId];
 
     return newState;
 }
